removed visualize option from AddEvent
events cant be visualized until after modules start, so in order for
the visualize param to work BCI2000Remote would have to keep track of
which evs need to be visualized, which is more trouble than it's worth
as we want to minimize internal state. Instead, use the Visualize()
method.