Reverted change to network status behavior
Reverted the change made in dac2148.

The more comprehensive connection check introduced 8ms of extra time
taken for each call, as opposed to the usual microsecond-scale latency
on a local connection. I have changed back to the native method for now,
as the result of attempting to send a command while not connected is the
same regardless of whether or not a proactive check is done, as in, an
exception is still thrown and must be handled by the caller.